[FIX] sap.f.DynamicPage: Fixed failure to expand the header
Problem:
The page fails to show the header even after it calls
_expandHeader to expand the header on resize (after
_shouldAutoExpandHeaderOnResize returned true).

This happens when the header still keeps a CSS class
sapFDynamicPageHeaderHidden that was set earlier by
another resize listener that happened to be called
first by the ResizeHandler. Thus the CSS class
sapFDynamicPageHeaderHidden forces the header to
remain hidden even after _expandHeader is called.

Solution:
Ensure that upon expanding the header, the
sapFDynamicPageHeaderHidden CSS class is
also removed.

Note:
A solution to add the removeal of the class
sapFDynamicPageHeaderHidden inside the _expandHeader
function itself was also researched but discarded for now,
because:
- there is a separate _toggleHeaderVisibility function
which, however is documented to be executable only if
user-interaction preceded the toggling (not the current case)
- in the previous such case (Change-Id:
I0d6f8ad70524b271e7668c15fa02aa6e00d917a6) the update of the
visibility was again kept separate from the _expandHeader function
- as this change is for a ticket downport, a firther refactoring
may be risky and can instead be done separately for master branch
only.

SNOW: CS20240008794565
Change-Id: I6f78439c1ccbd2aabfbabd9bc840f09e73cc8e8a
(cherry picked from commit 5219a40d2949b26770cc618362d813da90452f9e)
CR-Id: 002075125800004734372024