[FIX] sap/m/Input: Respect the busy indicator as part of the input
The update of the suggestion list requires the input field
remain being the active element. However, when the input is
set to busy, the focus has been shifted to the corresponding busy
indicator, which is rendered inside the input's DOM.

This focus behaviour was introduced with
Ie06f714fe09e9e78068aeb6796b05ce2ca8d2b30 to enable assistive technologies
to announce the busy state of the control.

This change ensures the busy indicator is respected as part of the
input control while maintaining the required focus behavior for the
suggestion list to function correctly.

SNOW: 002075129400012355672024
Change-Id: Ie46160310634e4415d32b58508841642856f741a