[INTERNAL]sap.ui.rta: Annotation Plugin
Adds an Annotation Plugin to Key User Adaptation. This plugin reads
annotation actions defined on the control designtime metadata and uses
the annotation dialog to retrieve the change information.

A single action or multiple actions can be defined. The plugin will
create the corresponding menu items, similarly to the "settings" plugin.

When the annotation dialog is closed with a valid change, the plugin
will trigger a restart of RTA so the change can be applied.

JIRA: TEAM42-405
Change-Id: Ic64166fedd0dddecf1dbd61a238f92a83b0161a5