[INTERNAL] Table: Add group header row context menu
In a setup with GridTableType and odata.v4.TableDelegate, group header rows have a context menu with options to remove all group levels and the group level where the context menu is opened.

The implementation does no depend on the delegate, so the context menu could appear in other scenarios as well. But the inner GridTable only creates group header rows in combination with the odata.v4.TableDelegate.

Change-Id: Ifd468dfa03287792b777cd760510b9ffc243fe7b
Jira: CPOUIFTEAMB-2387