[INTERNAL][FIX] sap.ui.rta: Clear flex state on RTA.stop if allContexts is set
If a key user lacks certain roles to see all available variants, a new
flex data request needs to be triggered after exiting RTA to ensure that
the restricted variants are no longer shown.
Since the connectors solely rely on the allContext flag in the session
storage, it is necessary to check this condition again after clearing it
in the RTA exit flow.
In newer versions these checks are executed in the FlexState.initialize
directly, which cannot be downported due to a high risk of side effects.

Therefore this change explicitly clears the flex state in the scenario
where the allContext flag changes, thus triggering a fresh request the
next time the FlexState is initialized, which is done by the lifecycle
hooks after the soft reload.

SNOW: DINC0363578
CR-Id: 002075125800004891262024
Change-Id: Ia7d2734dd21386cdbc208157402e973f2838a280