[FIX] We've now fixed the issue with the preview load while back navigating to the file in preview dialog's carousel
The dialog's setTitle property was rerendering the whole carousel when filename was set during page change. With this fix we are configuring customheader aggregtion of the dialog with title control to avoid rerendering of the whole dialog on page change in carousel. Now, the entire dialog is not re-rendered and carousel content remains visible on previous page navigation.

SNOW: DINC0368231
Change-Id: Ifbee19155272c499f06412601dbeba338219737d