[FIX] sap.m.DraftIndicator: Fix cutting of text
Problem: Text of DraftIndicator is cut on the right.

Solution: Problem comes from a recent update in the DOM structure of
the Label - there is an additional div with class sapMLabelInner between
the spans with classes sapMLabel and sapMLabeTextWrapper. It makes
DraftIndicator's styles for overriding overflow property not aplicable.
The CSS selector is fixed now.

SNOW: DINC0326402
Change-Id: I4cd54d36dcc4eae04c51e38116456dcee43403af