[FEATURE] sap.m.Link: Reactive area size API introduced
A new enumeration property called "reactiveAreaMode"
is now added. It is designed to make links easier
to activate and helps meet the WCAG 2.2 Target Size requirement.
It is applicable only for the SAP Horizon themes.
The Reactive area size area is sufficiently large
to help users avoid accidentally selecting (clicking or tapping)
on unintended UI elements.
There are two possible values for this property:
"Inline" and "Overlay".
The default value is "Inline" and it implies no change
in the visualization and behavior of the interactive
elements.
The second option is "Overlay", which indicates that
the interactive element will be displayed as an overlay
on top of other interactive parts of the page. In this case
the height of the reactive element's area is increased.

JIRA: BGSOFUIBALKAN-8065
Change-Id: I082bd510e0ed6f70735db7bc9fafaf0dfcfd7f1d