[INTERNAL] DemoKit: Usage-tracking now logs country
Problem:
When the user changes the language from the
Demokit settings, the usage-tracking then
logs the updated language but does not log
the corresponding region.

This happens because the framework specifies
region only for languages with more than one
options for regions, t.e. the framework skips
the region if it is the single possible one
(as it is known by default).

Solution:
Define the default region for each supported
language and include it in the tracked info.

SNOW: DINC0383827
Change-Id: I85a5d3e447923442447c0470c9e2d0e6d47a38f0