[INTERNAL] sap.m.DynamicDateRange: Fix Last Christmas sample to parse correctly
Previously when trying to type 'Last Christmas' in the input field, the date was not parsed correctly which means the Custom Option for that range (25 Dec - 26 Dec, Prev year) was not suggested as an option in the suggestion list.

Now with this change when typing 'Last Christmas' the date is parsed correctly and the Custom Option is suggested as an option in the suggestion list.

JIRA: BGSOFUIBALKAN-8987

Change-Id: I5fff6b6576ec98c6bb188afbcdb79422952a79c9