[INTERNAL] core/qunit/generic: exclude "ToggleSpinButton", "DateTimePickerPopup", "_ColorPickerBox", "CustomMonthPicker" and "CustomYearPicker" from being tested
-ToggleSpinButton is defined within sap/m/TimePicker and not in a
module of its own. Therefore it needs to be excluded from the
generic test

-DateTimePickerPopup is defined within sap/m/DateTimePicker and not
in a module of its own. Therefore it needs to be excluded from the
generic test

-_ColorPickerBox is defined within sap/ui/unified/ColorPicker and not
in a module of its own. Therefore it needs to be excluded from the
generic test

-CustomMonthPicker is defined within sap/ui/unified/CustomMonthPicker
and not in a module of its own. Therefore it needs to be excluded
from the generic test

-CustomYearPicker is defined within sap/ui/unified/CustomMonthPicker
and not in a module of its own. Therefore it needs to be excluded
from the generic test

SNOW: DINC0390742
SNOW: DINC0390775
CR-Id: 002075125800000532072025
Change-Id: I5267cef7840ef63d157ddcef1863624317380e96
(cherry picked from commit 80470e925e766af8dfa47f17bac9c91923eca104)