[FIX] DemoKit: Add missing callback
Problem:
No callback for the event of closing the cookie preferences
dialog in the case when the dialog is opened from
the header menu.

Solution:
Open the dialog via the base method (that both
opens and listens for closing of the dialog)
that is already used for opening the dialog
from the other locations.

Change-Id: I7ced4b5aa06dd5d4c30e1234472370f148ef0e48