[FIX] sap.m.Input: preserve value after rerendering of password field
To avoid adding the value attribute directly to the input DOM reference, we should remove it from the renderer and manage its synchronization through JavaScript logic. However, since the input value is updated on focus or Enter key press, we require a "live value" property. This can be achieved under the assumption that the input field always uses renderer version >=2. When having the string rendering or a DOM flush is performed, the entire DOM is replaced, making detection within the control not technically possible.

The solution retains the current operational logic but removes the attribute during the after rendering hook to maintain functionality seamlessly.

Change-Id: Ib90206973955f7b1d72e62b648d643abf1f5b012
CR-Id: 002075125800000534032025
SNOW: DINC0376300
SNOW: DINC0382948