[FIX] MTable: item count wrong if $count used
If the the table binding $count is used (oData V4) the m.Table announces the number of all items, even if not all items are read. So this count should also be announced by typeahead in Field.

Note: the number announced by Typeahead can differ by 1 to the number announced by m.Table as m.Table also counts the header as one row.

Change-Id: I12943f851efeab83e76cf71d79225d090c24dddc
SNOW: DINC0391757