[FIX] DemoKit: Sequence of tracked events corrected
Problem:
The event for beginning of the tracking session is
logged with delay (compared to events for page-view)
if it happens to wait longer for the info to be logged.

Solution:
Ensure the info to be logged (fetched asynchronously)
is available before adding the logs sequesnce, so that
the logs can be added *synchronously*, in the expected order.

Change-Id: I70228ab6536ed1790650c79c32497a292ed4fbf8
SNOW: DINC0396191