[FIX] sap.m.ObjectHeader: is correctly displayed on large screens
Problem: The area with statuses and attributes was placed incorrectly
next to the title / number area and not below it on a large resolution screens.

Root cause: One of the divs holding the statuses and attributes didn't have the max width required by design. It was only set in some cases.

Solution: Set the max width class always.

SNOW: DINC0388077

Change-Id: Ibf2ff07d9724b0352d2814f306788893bf67d6b2