[INTERNAL][FIX] sap.ui.fl: Require opt-in for FlexObject.prototype.canBeCondensed
Previously, the FlexObject.prototype.canBeCondensed base method always
returned true which enabled changes for condensing by default.
During the condensing process, the getIdForCondensing method is called
on the instance which is not implemented in the base class though.
As a result the condensing process failed for FlexObject subclasses
which forgot to explicitly opt-out by overriding canBeCondensed, as was
the case here for the ControllerExtensionChange.
This change switches canBeCondensed to an opt-in approach for higher
resilience.

SNOW: DINC0395590
Change-Id: I494d3f704d78f1c7860397ae741b497037bb9b6e