[INTERNAL][FIX] FilterPanel: set focus on added factory item
The focus is no longer lost when adding a new Filter in the dialog. It will be set on the latest added factory item (originally implemented in 6159387)

This change needs to fix it in mdc, as the panel has been moved in a later release to sap.m.p13n.

Due to the nature of the binding/rendering updates of the m.List, the manual focus update needs to be triggered in a timeout to not be overruled again.

CR-Id: 002075125800000689372025
SNOW: CS20240008660313
Change-Id: If65069c9f88081b9f35b641f9cce0d400dc16197