[INTERNAL] DemoKit: Extend check for functional cookies
Problem:
The TrustArc cookies preferences dialog since recently shows
an extended list of functional cookies. Namely, 3 more domans
unrelated to AdobeAnalytics usage tracking ("youtube.com",
"www.youtube.com", "google.com") are added.
The Demokit code automatically treats then as related to usage
tracking with AdobeAnalytics, which is incorrect.

Solution:
Update the check for consent for AdobeAnalytics usage tracking
to ignore the above cookies (treat them as different category).

Additionalyy add a missing check in the listener for changes
in cookie preferences.

Note:
An alternative solution would be to precisely list the domains
of cookies for usage-tracking (i.e. use whiteList, rather than
ignoreList). Given, however, that the allowList would be a long
list (9 entries currently) that could fluctuate in the future,
prefered the approach to treat ALL functional cookies as
related to usage tracking, excluding [the few] ones that
we know are unrelated to usage trackng with AdobeAnalyics.

Change-Id: I9425929454b5cecde1b83a3ace5c605efdb844b4