[INTERNAL] Integration Cards: Remove min-height for cards
The min-height on card level is no longer needed,
as there is a default min-height for each content.

Additionally if a card has only a header,
the header must have a title (as it is required), which will put the height of the card to at least 3rems.
Therefore min-height is not really needed.

The benefit is that when we don't have a default min-height,
we open the card for more layouting possibilities.

CR-Id: 002075125800000833772025
Snippix: https://snippix.only.sap/#486011
SNOW: DINC0395694
Change-Id: Idc88b0146d86fe83546cf6a1f3519fd430104251
(cherry picked from commit 5722402079db1d9e58efdfa2715f33928d08796e)