[INTERNAL] Engine: prevent unnecesary diff
The FilterController creates unnecessary changes right now, in case a propertyInfo object with key "items" is configured. This change prevents the unnecessary change creation, as #getBeforeApply will ensure the proper change creation.

Through this constellation, unnecessary changes will be persisted for end users which results in an irreversible state for the user with filters he can not remove, although the configuration is not likely to happen, the issue should be resolved.

SNOW: DINC0393773
CR-Id: 002075125800000893012025
Change-Id: Icaa16f51cd90db8f09b0ecd73e9640f8476641d3