[INTERNAL] Introduce manifest first for libraries
Load the manifest of a library first, to load dependent
resources as early as possible. This meens early loading of
css and the library-preload for dependent libraries.

Each library must wait for the dependent lib manifests and their css
to be loaded before loading its own css. This ensures the correct css
loading order.

To avoid deadlocks due to cyclic lib dependencies we build up an
initiator chain to only chain the load promises for the
first load call of a library.

JIRA: CPOUI5FRAMEWORK-720
Change-Id: I83a981b54f19643eea0a553186425c99dd2ebc76