[INTERNAL][FIX] ui.Table: showNotificationPopoverAtIndex called for row that doesn't exist
When the service does not provide a count, the binding assumes a length which is sum of the currently available rows and the table threshold. If this length is higher than the configured selection limit and an end user tries to select all rows, a popover should be displayed on the last selected row. However, in some cases this row doesn't exist and the method should handle it properly.

SNOW: DINC0379532
Change-Id: I8602c4b1f127003e0b7bd021d706a3e62da583ec