[FIX] Demo Kit: 'Borrowed from' links fixed
With using FormattedText for building links inside SubApiDetail page, we are omitting them from the body event handler for global linkClickHandler of DocumentationRouter. That way the href is not aligned as SEO/NON-SEO link address and guides to a non-existing page.
Changing the control to sap.m.Link is solving this issue.

SNOW: DINC0400654
Change-Id: I5a761915d5a425281f1ca788d8dad19f2bf40430