[FIX] sap.f.DynamicPage: fix not read Title
Problem: When DynamicPageTitle has "heading" aggregation with a Title
control in it and there is not header (no snapping/expanding
functionality), Title text is not read out.

Solution: Implement aria-label on the header element (same way as
ObjectPageLayout), which gets its text from the Title control in the
"heading" aggregation, if present. It can also come from another
supported Title-like controls (Label, Text, FormattedText).
If not present - aria-label can come from landmark info.
This functionality supports nesting of the Title control in the "header"
aggregation in FlexBoxes (VBox/HBox), as a general solution. The need of
this functionality comes from DINC0306269;

SNOW: DINC0179232
Change-Id: Ib7d29c278eb33d2e1dfcb5bb5601c31f61585c52