[FIX] sap.ui.unified.FileUploader: Drag and drop interaction adjusted
Issue:
- The file path was written into the input field on file drop,
but the file wasn't selected and couldn't be uploaded.
This issue was only reproducible with compact form factor mode.

Solution:
- The mask element displayed over the input field had less height
than the input field. Thus making it possible for the users to
drop a file over the input field. This input field is used only
for display purposes so file drops aren't expected over it.
The mask element height is now adjusted so it covers the input field entirely.

SNOW: CS20250009262449
Change-Id: I9b4172e86b71dc704debf3fab7e68335f6e0ae70