[FIX] sap.ui.unified.Calendar: The calendar week configuration has been properly applied
When the property's value is not pre-set, the value in the configuration should be used.
However, this does not occur because when the value is null (the default value of the property),
the first value in the enumeration, which is Default, is considered.
Consequently, the locale settings are taken into account.

After this change, it is checked whether the property's value has been modified.
If it has not been modified, the value null is replaced with undefined during the
calculation of the week number through the formatter.

SNIPPIX: 776490
SNOW: CS20240008957538
Change-Id: Ia841bda88353ba17fad8ff76829789a4f9fe5272