[FIX] Test Starter: Improve sinon compatibility
This change ensures that the sinon version loaded via the test starter
is exported to the global scope, so that it can be accessed in the
sinon-qunit-bridge and within the tests.

This is relevant when loading alternative versions of the modules as
pure AMD modules, where no export to the global scope is done.

In addition, the sinon.config will only be set when the version of
sinon provides a sinon.defaultConfig object, which has been removed in
newer versions.

Change-Id: Ib0f732d004134f1b9c12a1cbac2ae456cad888f2