[INTERNAL] dnd: dragleave event does not include relatedTarget in Safari
 - If the user drags an item (e.g., file) outside the browser window,
the central D&D mechanism informs the interested DropInfo objects and
highlights the droppable areas. The last detectable event when dragging
out of the browser window is the dragleave event. If the relatedTarget
property is empty, we can infer that the user has exited the browser
window. However, it appears that Safari never provides the
relatedTarget parameter for the dragleave event. This causes us to mistakenly assume that the user left the browser on every dragleave event.
 - As a workaround, we wait for 100ms after the dragleave event. If any
drag event (e.g. dragover or dragenter) occurs within this timeframe, we cancel the timer. If no drag event is detected, we conclude that the drag session has ended and proceed accordingly.
 - Reported Safari issue: https://bugs.webkit.org/show_bug.cgi?id=66547

SNOW: CS20250009203827

Change-Id: I453935a36699c3839d703448c159be339d270f4b