[FEATURE] sap.m.TimePicker: Improvements based on user-testing now implemented
* The clock dial in 24-hours format is now refactored as a single
circle, containing 24 units. This replaces the former 24-hours clock
dial, which consisted of two circles — an inner and an outer one —
each containing 12 units.

* Selecting a unit in the clock dial is now supported for both types
of units - dots and numbers. Formerly this feature was available only
for numbers.

* Hovering over a unit in the clock dial is now supported for both
types of units - dots and numbers. Formerly this feature was
available only for numbers.

* Animated transitions are introduced when changing the clock's
visualization with a click or a tap. Animated transitions are not
supported for the equivalent keyboard actions.

JIRA: BGSOFUIBALKAN-9066
Change-Id: Ida73d8c8f024d638381cdf089274b2ba5e3f4e11