[INTERNAL] ListBase: Skip warning when deselecting invisible context
If the selection state of an ODataV4 context is changed, and this context is not related to a visible item, a warning is logged.

This warning only makes sense if the context is selected, indicating that this has no effect on the list selection. Deselecting the context is irrelevant and no warning is needed.

Change-Id: Idec0797bc20c1336b8b6f749d2b2341c4d2beb3a