[INTERNAL] Adapt bootstrap to '1Chunk' strategy
The library-preload contains all boot relevant modules as well as the
library.js. The remaining modules are contained in the library-content.js
file. Library-preload and library-content are loaded in parallel.

Each boot task now can provide a beforeReady function which is called
after the boot phases and before the ready promose is resolved and the
Init module is executed. This e.g allows the splashscreen to register
to the first rendering to hide the splash screen, after booting and
loading all necessary resources.

JIRA: CPOUI5FRAMEWORK-759
Change-Id: I6caf1dbd73a69ccc72b5e53aa6c75083fc604e71