[FIX] sap.f.SemanticPage: fix unstable QUnit test
Problem: SemanticPage is not destroyed and onAfterRendering is executed
twice, while the second time is executed during the next QUnit
execution, so that - as a result, the next QUnit fails, because there is
an assertion from another test executed while this testis running.

Solution: Just destroying the SemanticPage in the QUnit, so no second
invalidation and onAfterRendering is happening and no intermittent
assertions are executed while running another test.

SNOW: DINC0422774
CR-Id: 002075125900000130802025
Change-Id: Id8df78ea64c5980d1782879295f2c30962491cce