[FIX] sap.f.FlexibleColumnLayout: Fix failing QUnit test
Problem: QUnit is failing due to previous tests not destroying the
created FCL. As a result, in the failing test we listen for "resume"
calls of the ResizeHandler and check the count, but it appears to be
more than expected, because it is called for other FCLs from different
tests.

Solution: Make sure each FCL is destroyed when tests are done.

SNOW: DINC0422728
CR-Id: 002075125900000137002025
Change-Id: I6620348effc5e3ea8579ab9a333b259b53deb65a