[FEATURE] ListItemBase: List item actions
- ListItemBase provide actions aggregations with the type
ListItemAction which has icon, text, type(Custom/Edit/Delete) and
visible property.
 - ListBase provides itemActionCount to define the max actions to be
visible in the items (max value is 2). If there are more than 2 actions
are defined the overflow button is rendered to let the user access
further actions.
 - The ListBase also fires itemActionPress event and informs which
action and item has been pressed.
 - Since FeedListItem already has the actions aggregation, a base class
has been created to unify the type and use the "actions" as aggregation
name.
 - GroupHeaderListItem and FeedListItem documents that
ListItemActionBase type is not supported

Jira: CPOUIFTEAMB-2039

Change-Id: I2d6df86bfbe33108f2e3262c95f1f1f7f2386cf9