[FIX] sap.ui.core.ShortcutHintsMixin: shows correctly the tooltip on mouseover
Problem: When the control that has tooltip generated with ShortcutHintsMixin is inside a container that also has a tooltip on hover of the Control the tooltip of the container control is shown over the ShortcutHintsMixin tooltip.

Solution: Add native empty tooltip (title attribute) on hover of the Control and remove it on mouseout this suppress the container native tooltip.

SNOW: DINC0406243

Change-Id: Ia3b291502cd00799622d332cf1afc678f10fc2cc