[FIX] sap.m.Popover: Scrolling issue in popover with overflowing content is now resolved
Problem:
During position calculations of the popover, all its styles are temporarily removed. This can cause an overflow in the documentElement, which can lead to a vertical scroll bar being displayed.
The width of this scroll bar might influence the "left" position of the opener. Later, when the styles are set again, this scrollbar is removed, which changes the opener's "left" position.
Opener repositioning triggers another position calculation of the popover, which can lead to an infinite loop.

Solution:
Ensure that there is no temporary scrollbar during the position calculation of the popover.

Fixes: https://github.com/SAP/openui5/issues/4216
SNOW: DINC0426831
Change-Id: I076e718a9e18881b4a604ae8830d1b77df67b11d