[FIX] sap.uxap.ObjectPageLayout: Incorrect selectedSection prevented
Problem:
Incorrect scroll position due to incorrect "selectedSection"
assigned in the listener to the "scroll" event. This happens
when:
1)the "scroll" event is caused by resize of the content
but
2)the cached positions of the sections' DOM elements are not
updated yet following the above resize

Findings:
There is an existing handling that sets a 'dirty' flag to
the cached DOM positions when the sections are invalidated,
but the 'dirty' flag is not closely synchronized with the
actual rerendering of the invalidated controls.
In particular, the dirty flag is set early (as soon as the control
is invalidated), after which the control waits to be rerendered.
So it is possible that while the contol waits to be rerendered,
the page updates the cached DOM positions by a different trigger
and thus clears the 'dirty' flag ^before* the waiting control
was rerendered in the DOM.

Solution:
- Ensure "dirty" flag is raised at the proper time, i.e.
  *right before rerendering* of the invalidated (sub)sections
- Ensure "dirty" flag is now raised also when the headerContent
  is rerenered.
- In addition, fix the height of the anchorBar during rerendering
  of its content, to prevent incorrect layout calulations in the
  interval when the [temporarily] empty anchorBar looses its height.

Change-Id: I0c0487ea1eb9b0f92180cd47ec4f878bb8f7ff23
SNOW: DINC0307327