[FEATURE] sap.uxap.ObjectPageSectionBase: API for effective title level
Background: There is a major redesign of ObjectPageSection/SubSection's
header area. As part of the redesign, the headers of the Section and
SubSection should be merged into one, when there is only one SubSection
in a given Section.

Implementation: In order to merge header areas, now we need to show the
SubSection title and its "actions", instead of the Section title on one
line and SubSection's action on another line. Showing the SubSection
title in this case requires the titleLevel and titleStyle to be
adjusted accordingly so that they can resemble the ones of the Section.
A new API on ObjectPageSectionBase level is provided -
getEffectiveTitleLevel. It returns the currently shown title level of
the ObjectPageSubSection (either its own, set by the application, or the
level of the Section if headers are merged). New API can be used by FE
to adjust the title level of the Forms, placed in a single
ObjectPageSubSection.

JIRA: BGSOFUIPIRIN-6627
Change-Id: I927cc298551a190fb45d5bc2d8f9709d3722fbfe