[FIX] sap.uxap.BlockBase: Fix not showing "moreBlocks"
Problem: When ObjectPageSubSection has "moreBlocks" aggregation and user
presses the "Show More" button - more blocks are not shown.

Solution: when connectToModels function of BlockBase (placed in
"moreBlocks" aggregation of ObjectPageSubSection) is called, the
BlockBase still does not have a parent (it is not added in the Grid),
and the parent ObjectPageLayout cannot be found. As a result bindings
are not updated and view is not selected. However, we have the parent
ObjectPageSubSection saved in a private property, when the view was
initialized. In this case, when calling getObjectPageLayout function
we need to pass the saved reference to the parent ObjectPageSubSection,
so that we can find the parent ObjectPageLayout.

SNOW: DINC0435859
Change-Id: I31cc1d82b8b2e8d409293ef564a60d00ffccdd1a