[FIX] File rename will fire the itemRenamed event on completion & will not trigger model updates in UploadsetwithTable plugin
With this change Uploadsetwithtable plugin will not update the model with renamed value as it would trigger the patch update / batch request in ODATA models, which might restrict applications / consumers of the plugin to manage the request to update the value in their models accordingly. Instead the file rename feature continues to fire itemRenamed event to notify the consuemrs the successful file rename and let consumers update the model as per need.

SNOW: DINC0426221
Change-Id: If5f90581d4181d99b79fddc61eec2aa61b19a9e2