[FIX] sap.tnt.NavigationListItem: fix firing of press event in popover items
For NavigationListItems inside the popover that opens in collapsed mode of SideNavigation,
the propagation of the 'press' event was not correct, which is fixed and now attached listener for 'press'
correctly recognizes pressing the item in both expanded and collapsed.

The itemPressed event on NavigationList and SideNavigation is now simplified and renamed to itemPress.

For NavigationListMenuItem, the correct library is now present in the class.

The ToolPage sample data is updated to have distinguishable item names.

SNOW: DINC0437313
Change-Id: I20ceecdb551857dac6b06cb87b819e44bc3ad846