[FIX] sap.m.Breadcrumbs: Fix issue when in OFT's Popover
Problem: When Breadcrumbs get into the Popover of the OverflowToolbar
and user tries to open the Popover - an error is thrown and Popover is
misplaced.

Solution: Popover gets opened and then closed, due to the rerendering of
the OverflowToolbar. OverflowToolbar gets rerendered because Breacrumbs
fires an invalidation event to it - for min required width changed.
However, this event is not needed in this case. OverflowToolbar needs to
know what minimum space to dedicate for the Breadcrumbs when it is in
the Toolbar line (in order all calculations for available space to be
correct), but not when it is in the Popover. So now, Breadcrumbs does
not fire an invalidation event to the OverflowToolbar, when it is placed
in the Popover.

SNOW: DINC0440942
Change-Id: I5a99976af5f907aab8877b42d2f10a9e93c715aa