[INTERNAL] sap.ui.unified.Calendar: InitialFocusedDate property works properly
Issue:
- The "initialFocusedDate" property was used in the "onBeforeRendering"
hook, in order to place a focus over a particular date in the grid.
That behavior interferes with the date selection in the grid
as the selection causes invalidation.
Further more each interaction that caused invalidation would result
into a focus retrieval to the same date in the grid.

Change-Id: I9819f28e08166568eba0a5ca6cf1b25109812fef