[INTERNAL] sap.ui.fl: enable deactivation of changes
Deleting a change is not always possible (e.g. different layers / backends).
Due to this a different way to deactivate changes is needed.
This can now be achieved by a change using the changeType 'deactivateChanges',
with the to be deactivated change ids in the content.
When loading the FlexObjects, before adding them to the FlexState, both the
deactivation changes as well as the to be deactivated changes are removed.

Example scenario: When creating an annotation change, any control based
change that will interfere with the annotation change needs to be deleted.

JIRA: TEAM42-426
Change-Id: Ida3c1db81218ec1f2e6d2e07a28a06c47422b8c2