[INTERNAL][FIX]sap.ui.fl: Fix VM usage of condense route
When VM changes are done in different layers, we should use the condense
route as all the necessary operations are available there.

This change ensures that every layer has its own save request, ensuring
that a condense request is used for each layer.

This change also removes the restriction that the condense route should
only be used for multiple changes, as it also works for every operation
of single changes.

SNOW: DINC0441056
Change-Id: Ifb673aa83aeefa4e89dc68b342d345fd0b5def6d