[INTERNAL][FIX] ColumnMenu: Reusing quick actions in menus
QuickAction content is rendered via the AssociativeControl, so it is not destroyed. The DOM of the destroyed QuickAction still contains the DOM of its undestroyed content, which may be used somewhere else.
The control is removed from the control tree so that upon destruction its DOM is removed synchronously.

SNOW: DINC0437427
Change-Id: Ie5a6af2c1ea6c82b4f1f447c60fb77fa74efddba