[INTERNAL] sap.ui.rta: introduce new properties in the annotation designtime
The delegate can now return two new properties per entry: label and tooltip.
Previously the propertyName was used as a label and there was no tooltip.
Now, if 'label' is provided, it will be displayed in the dialog; otherwise, 'propertyName' will be used as a fallback.
The tooltip is optional and there is no fallback.

JIRA: TEAM42-430
Change-Id: Icadf356179573ecf33323f5c9518b431f1b9833e