[FIX] Card Explorer: Events from previous card sample are no longer displayed on the current
Problem:
When navigating from one card sample to another, the events from the previous card sample are still displayed on the current card sample.

Reason:
setManifestChanges([]) is called on _showSample, then mulptiple invalidations of the card occur. This triggers the manifest to be applied again, before the new manifest is set.

Solution:
Along with the setManifestChanges([]), we need to also call setManifest(null). Refactored the code, so that _initCard() always resets the manifest and setManifestChanges([]) is called after it.
Additionally deleted a lot of dead code.

SNOW: DINC0143659
Change-Id: I38dedf3b4c25f11a2ce629f2513fb443a151831f