[INTERNAL] odata.v4.tableDelegate: Allow unit as CDP and fix text handling
Feature: Properties that are technically groupable are allowed to reference the unit in both the 'unit' and 'additionalProperties' attribute.

Fix: Context-defining properties do no longer override properties already added to $$aggregation.group. The issue was disappearing text properties, because the same property added was added later as a CDP with an empty object (missing group.additionally).

Change-Id: I8ff39e92e352e2d2db3a6c4a7f7c53d0e2c22218
Jira: CPOUIFTEAMB-2482
SNOW: DINC0447560