[INTERNAL] odata.v4.TableDelegate: Disable data aggregation plugin if it is not needed
If the plugin is enabled, it sets the GridTable to a rendering mode intended for a grouped structure. In this mode, the row selector cells are alwasy rendered, even if selection is disabled. This is not intended for simple lists without grouping or totals. So in this case, the plugin is disabled which sets the rendering mode of the GridTable back to normal.

Change-Id: Ia80b7ae91aeef7574bbef2fdfa8caa116e2cc0dd
SNOW: DINC0392015