[FIX] We now use messageBundle to set the text for the Upload button
UploadSetwithTable makes use of FileUploader to display the UploadButton. The button is set with the default text of "Upload".
The current text was hardcoded and which caused the same text "Upload" to be displayed when the control was used in other languages.

We have now updated the implementation to fetch display text from the messagebundle.properties which would be dynamically updated based on the language.

SNOW: DINC0453911
CR-Id: 002075125900000881832025
Change-Id: Ie555b4d1a9976f658e72a52c24884680b93dadbc
(cherry picked from commit 1d453c50d034cc1c278f551616b415c7a366c207)