[FIX] Field: opening dropdown selects first item
Problem:
If the user opens the dropdown via mouse or the dropdown opens automattically on focussing the Field, the first item is selected and the text is filled into the Field.
This should only happen if the user opens the dropdown via keyboard. (See behaviour of sap.m.ComboBox)

Solution:
If dropdown opens via mouse or automatically on focus the visual focus stays in the Field. No item is shown as selected in dropdown.
If dropdown opened via keyboad the visual focus moves to the dropdown. The fist item is shown as selected and focussed. The text of the item is transfered to the field.

Change-Id: Ia1240d41fba942d5d79120c244c0ea4a6163b890
SNOW: DINC0451198