[INTERNAL] sap.uxap.ObjectPageLayout: Remove scrollbar from tab chain
Problem: 
The scrollbar is focusable and can be reached by tabbing through the page. This can lead to confusion for users who expect to navigate through interactive elements only.

Solution:
To prevent the scrollbar from being focusable, we can use the tabindex attribute. By setting the tab index to -1, we can ensure that the scrollbar is skipped when tabbing through the page. This allows users to focus on the interactive elements without being distracted by the scrollbar.

Impact:
This change will improve the user experience for keyboard users by allowing them to navigate through the interactive elements of the page without being interrupted by the scrollbar.

Risk:
The risk of this change is low, as it only affects the focusability of the scrollbar. It does not change any functionality or behavior of the page. The module itself is only used by the ObjectPage.

Change-Id: If7faa0b12c9a29dc1ca382d4d94394680af59081
SNOW: CS20250009451759
CR-Id: 002075125900001129622025
(cherry picked from commit 3ee53c3271663daab386943f9e3d47a8edc0b4cf)