[INTERNAL] Input controls: Fix value state header links interactions
This change clears the problems found for keyboard interactions.

This change also contains a small fix related to synchronizations when closing suggestions dialog with value state header on mobile. The bug was that when closing the dialog, a value state message popup is opened and remains open on mobile devices.

Change-Id: I07fa2b20566cdbc30c11b129d2292002f8a1a86d