[INTERNAL]sap.ui.dt: Enable propagation of actions on DT Metadata
It is now possible to define actions that should be propagated to
children in the designtime metadata. This can be used to
propagate actions from a parent element to its children - like showing
the actions of a smarttable on a control inside it (e.g. the table).

This is done by having the action on the parent element and defining
the action names as strings
propagateActions: ["action1", "action2"]

or as an object with the action name and a function that returns
a boolean indicating if the action should be propagated to the child
propagateActions: [{action: "action1", isActive: (oChildElement) => ...}]

JIRA: TEAM42-434
Change-Id: Id75433ea912412978c17f250839ac9f3455db5b1